requirements.txt
flask>=2.0.0
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
pillow>=10.0.0
albumentations>=1.3.1
rarfile>=4.0
segment-anything>=1.0
numpy>=1.24.0
python-dotenv>=1.0.0
werkzeug>=2.0.0 
